lab3

+++ Lab3CPP/lab3.cpp
@@ -88,7 +88,7 @@ void bfs(Graph& graph, const string& start, const string& end, string* cityNames
             break;
         }
 
-        for (int i = 0; i < vector_size(graph.adjList[currentIdx]); ++i) {
+        for (int i = 0; i < static_cast<int>(vector_size(graph.adjList[currentIdx])); ++i) {
             int neighborIdx = vector_get(graph.adjList[currentIdx], i);
             if (vector_get(visited, neighborIdx) == 0) {
                 vector_set(visited, neighborIdx, 1);
